feat: implement proxy storage management and expiration checks
