+++ src/effect.ts
@@ -1,0 +1,60 @@
+import type { Effect, EffectFn, EffectMap } from '@/types'
+
+const storageEffectMap = new WeakMap<object, EffectMap>()
+
+export function on(
+    this: any,
+    storage: object,
+    key: string,
+    fn: EffectFn,
+) {
+    const effect: Effect = {
+        ctx: this,
+        fn,
+    }
+
+    let effectMap = storageEffectMap.get(storage)
+    if (!effectMap)
+        storageEffectMap.set(storage, (effectMap = new Map()))
+
+    const effects: Effect[] | undefined = effectMap.get(key)
+    if (effects)
+        effects.push(effect)
+    else
+        effectMap.set(key, [effect] as Effect[])
+}
+
+export function off(
+    storage: object,
+    key?: string,
+    fn?: EffectFn,
+) {
+    if (key === undefined) {
+        storageEffectMap.set(storage, new Map())
+        return
+    }
+
+    const effectMap: EffectMap | undefined = storageEffectMap.get(storage)
+    if (effectMap) {
+        const effects: Effect[] | undefined = effectMap.get(key)
+        if (effects && effects.length > 0) {
+            const value: Effect[] = fn ? effects.filter(ele => !(ele.fn === fn || (ele as any).fn?.fn === fn)) : []
+
+            effectMap.set(key, value)
+        }
+    }
+}
+
+export function trigger(
+    storage: object,
+    key: string,
+    value: any,
+    oldValue: any,
+) {
+    const effectMap: EffectMap | undefined = storageEffectMap.get(storage)
+    if (effectMap) {
+        const effects: Effect[] | undefined = effectMap.get(key)
+        if (effects)
+            effects.forEach(ele => ele.fn.call(ele.ctx, value, oldValue))
+    }
+}
