+++ src/extends/expires.ts
@@ -1,6 +1,7 @@
 import type { ExpiresType, StorageObject, StorageOptions } from '@/types'
 import { encode } from '@/proxy/transform'
-import { deleteProxyStorageProperty, getProxyStorageProperty } from '@/shared'
+import { deleteProxyStorageProperty, setProxyStorageProperty } from '@/cache'
+import { getProxyStorageProperty } from '@/shared'
 import { formatTime, isObject, pThen } from '@/utils'
 import { getOptions } from './options'
 
@@ -19,7 +20,9 @@ export function setExpires(
   pThen(() => getProxyStorageProperty(storage, property), (res: StorageObject | string | null) => {
     if (isObject(res)) {
       const options = Object.assign({}, res?.options, { expires: time })
-      const encodeValue = encode({ data: res.value, storage, property, options })
+      const newItem = { ...res, options }
+      setProxyStorageProperty(storage, property, newItem)
+      const encodeValue = encode({ data: res.value, options })
       storage.setItem(property, encodeValue)
     }
   })
@@ -44,30 +47,11 @@ export function removeExpires(
   pThen(() => getProxyStorageProperty(storage, property), (res: StorageObject | string | null) => {
     if (isObject(res) && res.options) {
       delete res.options.expires
-      const encodeValue = encode({ data: res.value, storage, property, options: res.options })
+      const newItem = { ...res, options: res.options }
+      setProxyStorageProperty(storage, property, newItem)
+      const encodeValue = encode({ data: res.value, options: res.options })
       storage.setItem(property, encodeValue)
     }
   })
 }
 
-export function checkExpired({
-  data,
-  storage,
-  property,
-}: {
-  data: StorageObject | string | null
-  storage: Record<string, any>
-  property: string
-}) {
-  if (!isObject(data) || !data.options)
-    return data
-
-  const { expires } = data.options
-
-  if (expires && new Date(+expires).getTime() <= Date.now()) {
-    deleteProxyStorageProperty(storage, property)
-    data.value = undefined
-  }
-
-  return data
-}
