+++ src/extends/watch.ts
@@ -1,30 +1,9 @@
-import type { Effect, EffectFn, EffectMap } from '@/types'
+import type { EffectFn } from '@/types'
+import { on, off, trigger } from '@/effect'
 import { postMessage } from '@/proxy/broadcast'
 import { hasChanged } from '@/utils'
 
-const storageEffectMap = new WeakMap<object, EffectMap>()
-
-export function on(
-  this: any,
-  storage: object,
-  key: string,
-  fn: EffectFn,
-) {
-  const effect: Effect = {
-    ctx: this,
-    fn,
-  }
-
-  let effectMap = storageEffectMap.get(storage)
-  if (!effectMap)
-    storageEffectMap.set(storage, (effectMap = new Map()))
-
-  const effects: Effect[] | undefined = effectMap.get(key)
-  if (effects)
-    effects.push(effect)
-  else
-    effectMap.set(key, [effect] as Effect[])
-}
+export { off, on, trigger } from '@/effect'
 
 export function once(
   this: any,
@@ -41,27 +20,6 @@ export function once(
   on(storage, key, wrapped)
 }
 
-export function off(
-  storage: object,
-  key?: string,
-  fn?: EffectFn,
-) {
-  if (key === undefined) {
-    storageEffectMap.set(storage, new Map())
-    return
-  }
-
-  const effectMap: EffectMap | undefined = storageEffectMap.get(storage)
-  if (effectMap) {
-    const effects: Effect[] | undefined = effectMap.get(key)
-    if (effects && effects.length > 0) {
-      const value: Effect[] = fn ? effects.filter(ele => !(ele.fn === fn || (ele as any).fn?.fn === fn)) : []
-
-      effectMap.set(key, value)
-    }
-  }
-}
-
 export function emit(
   storage: object,
   key: string,
@@ -74,18 +32,4 @@ export function emit(
 
   trigger(storage, key, value, oldValue)
   postMessage(storage, key, value, oldValue, property)
-}
-
-export function trigger(
-  storage: object,
-  key: string,
-  value: any,
-  oldValue: any,
-) {
-  const effectMap: EffectMap | undefined = storageEffectMap.get(storage)
-  if (effectMap) {
-    const effects: Effect[] | undefined = effectMap.get(key)
-    if (effects)
-      effects.forEach(ele => ele.fn.call(ele.ctx, value, oldValue))
-  }
 }
