+++ src/proxy/broadcast.ts
@@ -1,25 +1,28 @@
 import type { StorageLike } from '@/types'
-import { trigger } from '@/extends/watch'
-import { decode, simpleDecode, simpleEncode } from '@/proxy/transform'
-import { storageNameMap } from '@/shared'
+import { trigger } from '@/effect'
+import { decode, encode } from '@/proxy/transform'
+import { storageNameMap } from '@/cache'
 import { pThen } from '@/utils'
 
 const storageChannelMap = new Map<string, BroadcastChannel>()
 
+// Broadcast changes to other tabs/windows
 export function postMessage(storage: object, key: string, value: any, oldValue: any, property?: string) {
   const storageName = storageNameMap.get(storage)
   if (!storageName)
     return
 
   const channel = storageChannelMap.get(storageName)
+  console.log('postMessage', key, storageName)
   channel?.postMessage({
     key,
-    newValue: simpleEncode(value),
-    oldValue: simpleEncode(oldValue),
+    newValue: encode({ data: value }),
+    oldValue: encode({ data: oldValue }),
     property,
   })
 }
 
+// Listen for changes from other tabs/windows
 export function listenMessage(storage: StorageLike) {
   const storageName = storageNameMap.get(storage)!
 
@@ -29,7 +32,10 @@ export function listenMessage(storage: StorageLike) {
 
   channel.onmessage = function (ev: MessageEvent) {
     const { key, newValue, oldValue, property } = ev.data
-    trigger(storage, key, simpleDecode(newValue), simpleDecode(oldValue))
+    const newDecoded = decode({ data: newValue })
+    const oldDecoded = decode({ data: oldValue })
+
+    trigger(storage, key, newDecoded?.value, oldDecoded?.value)
 
     // update proxyStorage
     if (property) {
